separate the handler and uesr struct
