+++ interval/domain/user.go
@@ -1,1 +1,6 @@
 package domain
+
+type User struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
